requirements.txt
streamlit == 1.48.0
streamlit-webrtc
fastapi
uvicorn
chromadb
langchain-community 
sqlite3
langchain-google-genai
langchain
langchain-text-splitters
python-dotenv
google-generativeai